items: reused existing templates and css classes to display product pictures
My own personnal rant: wells are a pain in the ass to use, meaning very hard to manipulate within pages relying a lot on grid (like ours), for a 'styling' gain that is highly questionable.
So, from now on, let's keep wells use for plain text (no columns), whichs means wrap the well within a span-n, and not put them at the same level.